router/api: moves fake impl to different file
